geanygendoc: Hide a harmless GCC warning
GTokenType don't have named constants for every possible character, so
GCC will warn for byte values not explicitly part of that enumeration,
even though it has a gap designed for any byte to be valid.

+++ geanygendoc/src/ggd-file-type-loader.c
@@ -408,7 +408,7 @@ ggd_file_type_read_match (GScanner  *scanner,
         
         identifier = g_strdup (scanner->value.v_identifier);
         /*g_debug ("identifier is %s", identifier);*/
-        switch (g_scanner_get_next_token (scanner)) {
+        switch ((guint) g_scanner_get_next_token (scanner)) {
           case '.': /* skip it */ break;
           
           case G_TOKEN_EQUAL_SIGN:
